refactor: change payload factory and response processor to pipelines
