+++ src/sc_client/core/sc_connection.py
